migrating deckgl chart
